requirements.txt
pandas>=2.0
numpy>=1.24
scipy>=1.10
statsmodels>=0.14
lifelines>=0.27
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
statsbombpy>=1.11
requests>=2.28
beautifulsoup4>=4.12
pyarrow>=12.0
tqdm>=4.65
pyyaml>=6.0
jupyter>=1.0
notebook>=7.0
